Added the media field to TypoherbariumObservation.
Added the TypoherbariumMedia class to the typoherbarium model.

Added the ORM between the TypoherbariumMedia class and the iherba_medias table
as as set of methods of the PersistentLocalTypoherbarium class.

Modified the observationReceive.php to be able to receive medias.

Tested, that all that does not break other things. However did not
really test if it works.